Structure the tui by responsibility
